+++ HISTORY.rst
@@ -1,6 +1,13 @@
 =======
 History
 =======
+1.2.0 (2019-11-30)
+------------------
+* Add LeavingWaterTemp and WaterFlowRate sensors (Series 7 WF)
+* Added features to CLI (for details see --help)
+*   Added debug flag 
+*   Added option sensor list specification 
+*   Added continous reporting 
 
 1.1.0 (2019-01-07)
 ------------------